+++ tests/test_wordalign_filter.py
@@ -6,7 +6,7 @@ import os
 import tempfile
 import unittest
 
-from opusfilter import word_alignment
+from opusfilter import word_alignment, OpusFilterRuntimeError
 
 
 @unittest.skipIf(os.environ.get('EFLOMAL_PATH') is None, 'EFLOMAL_PATH not defined in environment')
@@ -61,10 +61,15 @@ class TestAlignFilter(unittest.TestCase):
         """
         prior_data1 = ['%s.' % ('ab ' * (line + 1)) for line in range(10)] * 5
         prior_data2 = ['%s.' % ('AB ' * (line + 1)) for line in range(10)] * 5
-        priors_file = tempfile.NamedTemporaryFile('w+')
-        word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name)
-        self.assertTrue(priors_file.read())
-        priors_file.close()
+        with tempfile.NamedTemporaryFile('w+') as priors_file:
+            word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name)
+            self.assertTrue(priors_file.read())
+
+    def test_make_priors_empty(self):
+        """Test making priors with empty data"""
+        with self.assertRaises(OpusFilterRuntimeError):
+            with tempfile.NamedTemporaryFile('w+') as priors_file:
+                word_alignment.make_priors([], priors_file.name)
 
     def test_make_priors_with_scores(self):
         """Test making priors with the score file option
@@ -75,14 +80,11 @@ class TestAlignFilter(unittest.TestCase):
         """
         prior_data1 = ['%s.' % ('ab ' * (line + 1)) for line in range(10)] * 5
         prior_data2 = ['%s.' % ('AB ' * (line + 1)) for line in range(10)] * 5
-        priors_file = tempfile.NamedTemporaryFile('w+')
-        score_file = tempfile.NamedTemporaryFile('w+')
-        word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name, score_file=score_file.name)
-        self.assertTrue(priors_file.read())
-        scores = [json.loads(line) for line in score_file]
-        self.assertEqual(len(scores), len(prior_data1))
-        priors_file.close()
-        score_file.close()
+        with tempfile.NamedTemporaryFile('w+') as priors_file, tempfile.NamedTemporaryFile('w+') as score_file:
+            word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name, score_file=score_file.name)
+            self.assertTrue(priors_file.read())
+            scores = [json.loads(line) for line in score_file]
+            self.assertEqual(len(scores), len(prior_data1))
 
     def test_scoring_priors(self):
         """Test word alignment scoring on artificial data using priors
@@ -95,17 +97,16 @@ class TestAlignFilter(unittest.TestCase):
         prior_data2 = ['%s.' % ('AB ' * (line + 1)) for line in range(10)] * 5
         data1 = ['%s.' % ('ab ' * (line + 1)) for line in range(5)] + ['', 'ab ab ab .']
         data2 = ['%s.' % ('AB ' * (line + 1)) for line in range(5)] + ['', 'AB']
-        priors_file = tempfile.NamedTemporaryFile('w+')
-        word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name)
-        align_filter = word_alignment.WordAlignFilter(src_threshold=0, tgt_threshold=0, priors=priors_file.name)
-        scores = []
-        bools = []
-        for score in align_filter.score(zip(data1, data2)):
-            scores.append(score)
-            bools.append(align_filter.accept(score))
-        logging.info(scores)
-        self.assertSequenceEqual(bools, [True] * 5 + [True, False])
-        priors_file.close()
+        with tempfile.NamedTemporaryFile('w+') as priors_file:
+            word_alignment.make_priors(zip(prior_data1, prior_data2), priors_file.name)
+            align_filter = word_alignment.WordAlignFilter(src_threshold=0, tgt_threshold=0, priors=priors_file.name)
+            scores = []
+            bools = []
+            for score in align_filter.score(zip(data1, data2)):
+                scores.append(score)
+                bools.append(align_filter.accept(score))
+            logging.info(scores)
+            self.assertSequenceEqual(bools, [True] * 5 + [True, False])
 
     def test_filtering(self):
         """Test word alignment filtering on artificial data